Added __bindEvents function for autogenerate event bindings on Ply Views
Now, when defining a new Ply view, the user can define an `__events` objects on the prototype, that takes as its keys an event name and the object id, separated by a space, e.g. `'click closeBtn'` would be used to attach a click handler to `this.objects.closeBtn`. As its values, it takes the name of the function to execute, e.g. `'close'` to run `this.close`. The callback function takes two arguments: the first is the event object, and the second is a shortcut to the element that triggered the event. Within the function, `this` refers to the Ply object, not the element, like in previous versions of Ply. This removes the need to use `self`, and allows for referencing other methods.